Add the concept of default dimensions
Dimensions can not be a dictionary since that will break with colelctors
that do not ahve dimensions in their base class. if self.dimensions is
None *as it is for other versions of diamond* doing
self.dimensions[dim_name] will break since None is not a dict obviously.

This commit also handles collector level and metric level dimensions
as is the case as well with the go collectors