Prefixing collectors and dimensioanlising them
Adding some of the metric names to the main namespace is a very bad idea
for example the EntropyCollector had the metric => 'available' in the
main name space that is really frustrating.

Also since we now have support for dimensions, I have incorporated
dimensionalising the collectors that have been recently newly migrated from
Ganglia and those that existed but have been touched due to the
migraiton.