Suffice with notifying when a collector over runs
There is no sane way to kill a g routine or stop exection of a function
in go that I was able to find. Hence I will suffice with the defer,
panic, recover technique using syscalls to be able to report when such
misbehaviour happens with a collector. This will not kill the colleciton
however will report this incident.

As per go 1.2 any go routine taking *too much* cpu time as determined by
the runtime scheduler will be yielded and will stop processing so that
should take care of itself.

I have also ensred that only one colleciton takes places at a time by
blocking in the collect ticker channel.