Errors monitoring for collectors
Montoring the state of your collectors is very important if you don't
want to fill up your logs with errors undetected. It is also very useful
to see where , when and why your collectors fail. We have found cases
where a collection fails only on a subset of hosts, since they had a bad
seyup, while it works well on the other 90%. Having this sort of
monitoring adds a layer of resilience to your collections and gives you
a more hollistic view of your systems.