Added scribe handler
The new scribe handler writes the metrics to a scribe stream as JSON
objects. All the metrics are written to the same stream to avoid
creating countless small streams.

Since we don't want all the metrics to be written to scribe, this
handler must always be configured with a whitelist. This way only the
metrics coming from the whitelisted collectors will be written to
scribe.