Make the batch size configurable
We have seen some collectors that are capable of sending hige messages
to the fullerite core, which causes lots of errors and also causes the
collector process to hold a lot in memory at a time. For such collectors
we can tweak the max_buffer_size config variable 'defaulted to 300'
to a value we see fit.